# Seperate trie commit and hash