--- a/storage/statedb/committer.go
+++ b/storage/statedb/committer.go
@@ -46,7 +46,7 @@ type leaf struct {
 // processed sequentially - onleaf will never be called in parallel or out of order.
 type committer struct {
 	tmp sliceBuffer
-	sha keccakState
+	sha KeccakState
 
 	onleaf LeafCallback
 	leafCh chan *leaf
@@ -57,7 +57,7 @@ var committerPool = sync.Pool{
 	New: func() interface{} {
 		return &committer{
 			tmp: make(sliceBuffer, 0, 550), // cap is as large as a full fullNode.
-			sha: sha3.NewLegacyKeccak256().(keccakState),
+			sha: sha3.NewLegacyKeccak256().(KeccakState),
 		}
 	},
 }