--- a/storage/statedb/committer.go
+++ b/storage/statedb/committer.go
@@ -1,3 +1,4 @@
+// Modifications Copyright 2022 The klaytn Authors
 // Copyright 2019 The go-ethereum Authors
 // This file is part of the go-ethereum library.
 //
@@ -13,6 +14,9 @@
 //
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
+//
+// This file is derived from trie/comitter.go (2022/11/14).
+// Modified and improved for the klaytn development.
 
 package statedb