--- a/storage/statedb/hasher.go
+++ b/storage/statedb/hasher.go
@@ -28,10 +28,10 @@ import (
 	"golang.org/x/crypto/sha3"
 )
 
-// keccakState wraps sha3.state. In addition to the usual hash methods, it also supports
+// KeccakState wraps sha3.state. In addition to the usual hash methods, it also supports
 // Read to get a variable amount of data from the hash state. Read is faster than Sum
 // because it doesn't copy the internal state, but also modifies the internal state.
-type keccakState interface {
+type KeccakState interface {
 	hash.Hash
 	Read([]byte) (int, error)
 }
@@ -50,7 +50,7 @@ func (b *sliceBuffer) Reset() {
 // hasher is a type used for the trie Hash operation. A hasher has some
 // internal preallocated temp space
 type hasher struct {
-	sha keccakState
+	sha KeccakState
 	tmp sliceBuffer
 }
 
@@ -59,7 +59,7 @@ var hasherPool = sync.Pool{
 	New: func() interface{} {
 		return &hasher{
 			tmp: make(sliceBuffer, 0, 550), // cap is as large as a full fullNode.
-			sha: sha3.NewLegacyKeccak256().(keccakState),
+			sha: sha3.NewLegacyKeccak256().(KeccakState),
 		}
 	},
 }