@hqjang-pepper 2022-11-11
### Benchmark Hash and commit



- #### Hash

```
Before
BenchmarkHashFixedSize
BenchmarkHashFixedSize/10
BenchmarkHashFixedSize/10-10 	           27526	     43064 ns/op	   14005 B/op	     157 allocs/op
BenchmarkHashFixedSize/100
BenchmarkHashFixedSize/100-10         	   10000	    101392 ns/op	   65680 B/op	     777 allocs/op
BenchmarkHashFixedSize/1K
BenchmarkHashFixedSize/1K-10          	    1190	   1004009 ns/op	  657317 B/op	    7782 allocs/op
BenchmarkHashFixedSize/10K
BenchmarkHashFixedSize/10K-10         	     274	   4429889 ns/op	 6905728 B/op	   79660 allocs/op
BenchmarkHashFixedSize/100K
BenchmarkHashFixedSize/100K-10        	      27	  38932048 ns/op	68239165 B/op	  794500 allocs/op

After
BenchmarkHashFixedSize
BenchmarkHashFixedSize/10
BenchmarkHashFixedSize/10-10 	           37155	     32239 ns/op	   11942 B/op	     135 allocs/op
BenchmarkHashFixedSize/100
BenchmarkHashFixedSize/100-10         	    7766	    153061 ns/op	   57988 B/op	     671 allocs/op
BenchmarkHashFixedSize/1K
BenchmarkHashFixedSize/1K-10          	     798	   1505582 ns/op	  584060 B/op	    6728 allocs/op
BenchmarkHashFixedSize/10K
BenchmarkHashFixedSize/10K-10         	      66	  18371760 ns/op	 6224453 B/op	   69396 allocs/op
BenchmarkHashFixedSize/100K
BenchmarkHashFixedSize/100K-10        	       6	 177527841 ns/op	61796261 B/op	  693042 allocs/op
```


- #### Commit
```
Before
BenchmarkCommitAfterHashFixedSize
BenchmarkCommitAfterHashFixedSize/10
BenchmarkCommitAfterHashFixedSize/10-10 	           30988	     38790 ns/op	   36083 B/op	     208 allocs/op
BenchmarkCommitAfterHashFixedSize/100
BenchmarkCommitAfterHashFixedSize/100-10         	    8286	    145892 ns/op	  174370 B/op	    1022 allocs/op
BenchmarkCommitAfterHashFixedSize/1K
BenchmarkCommitAfterHashFixedSize/1K-10          	     828	   1485660 ns/op	 1701478 B/op	   10143 allocs/op
BenchmarkCommitAfterHashFixedSize/10K
BenchmarkCommitAfterHashFixedSize/10K-10         	      74	  14636495 ns/op	18696716 B/op	  103985 allocs/op
BenchmarkCommitAfterHashFixedSize/100K
BenchmarkCommitAfterHashFixedSize/100K-10        	       7	 148162095 ns/op	180172019 B/op	 1038887 allocs/op

After
BenchmarkCommitAfterHashFixedSize
BenchmarkCommitAfterHashFixedSize/10
BenchmarkCommitAfterHashFixedSize/10-10 	           60080	     20030 ns/op	   30914 B/op	     189 allocs/op
BenchmarkCommitAfterHashFixedSize/100
BenchmarkCommitAfterHashFixedSize/100-10         	   12500	     96729 ns/op	  151233 B/op	     918 allocs/op
BenchmarkCommitAfterHashFixedSize/1K
BenchmarkCommitAfterHashFixedSize/1K-10          	    1052	   1188825 ns/op	 1473359 B/op	    9108 allocs/op
BenchmarkCommitAfterHashFixedSize/10K
BenchmarkCommitAfterHashFixedSize/10K-10         	      94	  11857340 ns/op	16368251 B/op	   93862 allocs/op
BenchmarkCommitAfterHashFixedSize/100K
BenchmarkCommitAfterHashFixedSize/100K-10        	       9	 121488097 ns/op	157054739 B/op	  937567 allocs/op
```